Working AI Loop
The loops works decently well now, it wander around aimlessly until it
is hungry, and then try to find food to eat, problem is when it somehow
can't find a path, it freezes up entirely.

I think the freezing has to do with the target being stuck in a place
where the best path isn't possible because it will just go back and
forth (going right is closer and has a lower value, but from there, you
only have impassables to your right side, and the optimal path is now
going back.... you get the point), we might need to optimize that.